requirements.txt
streamlit==1.30.0
numpy>=1.22.0
pandas>=1.5.0
scikit-learn>=1.2.0
tensorflow>=2.10.0
pytesseract>=0.3.10
Pillow>=9.0.0
PyMuPDF>=1.20.0
python-docx>=0.8.10
joblib>=1.1.0
matplotlib>=3.5.0